Move content management to App
